Move some logic out of eleventy.config.ts for reuse by changelog script
